+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Migrate to null safety
 - Improve usability of API using relative values for `depth` and `velocity`.
+- Replace `LitStarfieldBackground` with `LitStarfieldContainer`
 
 ## 0.0.1